refactor(crawler): extract discovery module
